backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pandas>=2.0.0
loguru>=0.7.0
requests>=2.31.0
finnhub-python>=2.4.0
aiohttp>=3.9.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
pytest>=7.4.0
httpx>=0.25.0
pathlib2>=2.3.7